+++ src/services/badge.js
@@ -7,8 +7,8 @@ const PAUSE_LOGO = './images/logo-yellow.png'
 
 export default {
   stop(text) {
-    chrome.browserAction.setIcon({ path: DEFAULT_LOGO })
-    chrome.browserAction.setBadgeBackgroundColor({ color: DEFAULT_COLOR })
+    chrome.action.setIcon({ path: DEFAULT_LOGO })
+    chrome.action.setBadgeBackgroundColor({ color: DEFAULT_COLOR })
     this.setText(text)
   },
 
@@ -17,19 +17,19 @@ export default {
   },
 
   setText(text) {
-    chrome.browserAction.setBadgeText({ text })
+    chrome.action.setBadgeText({ text })
   },
 
   pause() {
-    chrome.browserAction.setIcon({ path: PAUSE_LOGO })
+    chrome.action.setIcon({ path: PAUSE_LOGO })
   },
 
   start() {
-    chrome.browserAction.setIcon({ path: RECORDING_LOGO })
+    chrome.action.setIcon({ path: RECORDING_LOGO })
   },
 
   wait() {
-    chrome.browserAction.setBadgeBackgroundColor({ color: RECORDING_COLOR })
+    chrome.action.setBadgeBackgroundColor({ color: RECORDING_COLOR })
     this.setText('wait')
   },
 }