+++ src/services/storage.js
@@ -4,7 +4,7 @@ export default {
       return Promise.reject('Browser storage not available')
     }
 
-    return new Promise(resolve => chrome.storage.local.get(keys, props => resolve(props)))
+    return new Promise((resolve) => chrome.storage.local.get(keys, (props) => resolve(props)))
   },
 
   set(props) {
@@ -12,7 +12,7 @@ export default {
       return Promise.reject('Browser storage not available')
     }
 
-    return new Promise(resolve => chrome.storage.local.set(props, res => resolve(res)))
+    return new Promise((resolve) => chrome.storage.local.set(props, (res) => resolve(res)))
   },
 
   remove(keys) {
@@ -20,6 +20,6 @@ export default {
       return Promise.reject('Browser storage not available')
     }
 
-    return new Promise(resolve => chrome.storage.local.remove(keys, res => resolve(res)))
+    return new Promise((resolve) => chrome.storage.local.remove(keys, (res) => resolve(res)))
   },
 }